Misc hacking

+++ whitenoise/gzip.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-from __future__ import absolute_import, print_function, division
+from __future__ import absolute_import, print_function, division, unicode_literals
 
 import argparse
 import gzip
@@ -9,26 +9,31 @@ import re
 
 CHUNK_SIZE = 64 * 1024
 
-def main(root, extensions, quiet=False):
+
+def main(root, extensions, quiet=False, force=False):
     file_re = re.compile(r'\.({})$'.format('|'.join(map(re.escape, extensions))))
     log = (lambda x:x) if quiet else print
     for dirpath, dirs, files in os.walk(root):
         for filename in files:
             if file_re.search(filename):
                 path = os.path.join(dirpath, filename)
-                compress(path, log)
+                compress(path, log, force)
 
-def compress(path, log):
+
+def compress(path, log, force):
     gzip_path = path + '.gz'
+    if not force and os.path.exists(gzip_path):
+        log('Skipping {} (.gz file already exists, use --force to overwrite)'.format(path))
+        return
     with open(path, 'rb') as in_file:
         with gzip.open(gzip_path, 'wb', compresslevel=9) as out_file:
-            for chunk in iter(lambda: in_file.read(CHUNK_SIZE), ''):
+            for chunk in iter(lambda: in_file.read(CHUNK_SIZE), b''):
                 out_file.write(chunk)
     # If gzipped file isn't actually any smaller then get rid of it
     orig_size = os.path.getsize(path)
     gzip_size = os.path.getsize(gzip_path)
     if gzip_size >= orig_size:
-        log('Skipping {} (No gzip saving)'.format(path))
+        log('Skipping {} (Gzip file is larger)'.format(path))
         os.unlink(gzip_path)
     else:
         log('Gzipping {} ({}K -> {}K)'.format(
@@ -41,7 +46,8 @@ if __name__ == '__main__':
                         "and produce gzipped versions with a '.gz' suffix (as long "
                         "this results in a smaller file.",
             formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-    parser.add_argument('--quiet', help="Don't produce log output", action='store_true')
+    parser.add_argument('-q', '--quiet', help="Don't produce log output", action='store_true')
+    parser.add_argument('-f', '--force', help="Overwrite pre-existing .gz files", action='store_true')
     parser.add_argument('root', help='Path root from which to search for files')
     parser.add_argument('extensions', nargs='*', help='File extensions to match', default=('css', 'js'))
     args = parser.parse_args()
